+++ src/icons/logos/Instagram.tsx
@@ -1,0 +1,185 @@
+// Import custom components.
+
+// Import custom utilities.
+
+// Import custom types.
+import { IconComponentProps } from '@/src/lib/types';
+
+// Import styles.
+
+/**
+ * This function renders an SVG icon of the Instagram logo.
+ */
+const Instagram = ({
+    color = '#000000',
+    height = '1em',
+    strokeWidth = 1,
+    width = '1em'
+}: IconComponentProps) => {
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            xmlSpace="preserve"
+            style={{
+                fillRule: 'evenodd',
+                clipRule: 'evenodd',
+                strokeLinejoin: 'round',
+                strokeMiterlimit: 2
+            }}
+            viewBox="0 0 32 33"
+            height={height}
+            width={width}
+            stroke={color}
+            strokeWidth={strokeWidth}
+            fill={color}
+        >
+            <ellipse
+                cx={621.368}
+                cy={81.007}
+                rx={0.649}
+                ry={0.642}
+                style={{
+                    fill: 'url(#ic_category_social_media_instagram_svg__a)'
+                }}
+                transform="translate(-1817.956 -232.194)scale(2.96649)"
+            />
+            <path
+                d="M621.368 80.365c.359 0 .649.288.649.642s-.29.643-.649.643a.646.646 0 0 1-.649-.643c0-.354.291-.642.649-.642m0 .457a.205.205 0 0 1 .214.185c.003.061-.007.12-.044.162-.046.051-.097.106-.17.112-.086.008-.171.004-.233-.046-.071-.058-.137-.13-.147-.228a.4.4 0 0 1 .073-.302.47.47 0 0 1 .307-.179z"
+                style={{
+                    fill: 'none'
+                }}
+                transform="matrix(2.96649 0 0 2.96648 -1817.953 -232.198)"
+            />
+            <path
+                d="m481.115 145.886.192-5.071c.048-1.474-.399-2.823-1.137-3.828-.758-1.032-1.837-1.672-3.04-1.73l-8.314-.254a3.8 3.8 0 0 0-1.773.423 4.7 4.7 0 0 0-1.51 1.256 6.1 6.1 0 0 0-1.008 1.896 7.1 7.1 0 0 0-.344 2.237l.135 10.37a6.9 6.9 0 0 0 .375 2.151c.233.667.565 1.28.992 1.793.423.509.915.908 1.45 1.169a3.6 3.6 0 0 0 1.683.375l8.314-.283c1.13-.041 2.143-.642 2.863-1.601.704-.939 1.131-2.211 1.099-3.604l-.107-4.417c0-.021-.014-.038-.031-.038-.016 0-.03.017-.03.038l-.107 4.417c-.029 1.267-.463 2.398-1.141 3.205-.663.788-1.564 1.255-2.546 1.22l-8.314-.283a2.7 2.7 0 0 1-1.245-.358 3.4 3.4 0 0 1-1.031-.919 4.3 4.3 0 0 1-.666-1.331 4.9 4.9 0 0 1-.209-1.534l.136-10.37c.002-.503.087-.993.238-1.448s.369-.873.651-1.229c.28-.357.611-.638.971-.831a2.5 2.5 0 0 1 1.155-.31l8.314-.254c.909-.044 1.748.369 2.368 1.091.642.747 1.065 1.794 1.103 2.981l.193 5.071c0 .112.072.202.16.202s.161-.09.161-.202"
+                style={{
+                    fill: 'url(#ic_category_social_media_instagram_svg__b)'
+                }}
+                transform="matrix(1.83017 0 0 1.45378 -849.25 -195.662)"
+            />
+            <path
+                d="M560.301 96.456c4.42.448 8.108 2.616 10.746 5.653 2.418 2.785 3.712 6.402 3.397 10.296-.138 3.679-1.953 6.853-4.491 9.249-2.517 2.376-5.844 4.031-9.652 4.107-3.937.177-7.663-1.122-10.431-3.454-2.9-2.445-4.875-5.865-4.94-9.822-.104-4.258 1.045-7.826 3.428-10.579 2.43-2.806 5.993-4.784 10.594-5.837a.277.277 0 0 0 .236-.317.287.287 0 0 0-.331-.226c-5.085.479-9.173 2.169-12.194 4.839-3.262 2.882-5.187 6.925-5.75 11.96-.547 5.287 1.534 10.163 5.05 13.815 3.564 3.701 8.725 5.96 14.338 6.213 5.785.116 11.153-1.988 14.92-5.652 3.751-3.648 5.901-8.778 5.694-14.296-.43-5.321-2.816-10.08-6.788-13.247-3.743-2.984-8.607-4.636-13.826-4.106-.405 0-.733.314-.733.702s.328.702.733.702"
+                style={{
+                    fill: 'url(#ic_category_social_media_instagram_svg__c)'
+                }}
+                transform="matrix(.2465 .31697 -.33081 .25727 -84.031 -189.87)"
+            />
+            <defs>
+                <radialGradient
+                    id="ic_category_social_media_instagram_svg__b"
+                    cx={0}
+                    cy={0}
+                    r={1}
+                    gradientTransform="matrix(-14.3873 19.2544 -15.2946 -18.1122 479.977 136.999)"
+                    gradientUnits="userSpaceOnUse"
+                >
+                    <stop
+                        offset={0}
+                        style={{
+                            stopColor: '#515bd4',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={0.3}
+                        style={{
+                            stopColor: '#8134af',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={0.55}
+                        style={{
+                            stopColor: '#dd2a7b',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={0.77}
+                        style={{
+                            stopColor: '#f58529',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={1}
+                        style={{
+                            stopColor: '#feda77',
+                            stopOpacity: 1
+                        }}
+                    />
+                </radialGradient>
+                <radialGradient
+                    id="ic_category_social_media_instagram_svg__c"
+                    cx={0}
+                    cy={0}
+                    r={1}
+                    gradientTransform="matrix(14.772 90.6031 -94.5601 14.1538 553.256 70.003)"
+                    gradientUnits="userSpaceOnUse"
+                >
+                    <stop
+                        offset={0}
+                        style={{
+                            stopColor: '#515bd4',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={0.3}
+                        style={{
+                            stopColor: '#8134af',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={0.55}
+                        style={{
+                            stopColor: '#dd2a7b',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={0.77}
+                        style={{
+                            stopColor: '#f58529',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={1}
+                        style={{
+                            stopColor: '#feda77',
+                            stopOpacity: 1
+                        }}
+                    />
+                </radialGradient>
+                <linearGradient
+                    id="ic_category_social_media_instagram_svg__a"
+                    x1={0}
+                    x2={1}
+                    y1={0}
+                    y2={0}
+                    gradientTransform="matrix(-1.95212 2.28653 -2.26205 -1.93122 622.41 79.714)"
+                    gradientUnits="userSpaceOnUse"
+                >
+                    <stop
+                        offset={0}
+                        style={{
+                            stopColor: '#515bd4',
+                            stopOpacity: 1
+                        }}
+                    />
+                    <stop
+                        offset={1}
+                        style={{
+                            stopColor: '#8134af',
+                            stopOpacity: 1
+                        }}
+                    />
+                </linearGradient>
+            </defs>
+        </svg>
+    );
+};
+
+export default Instagram;